backend/requirements.txt
# Flask核心
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
Flask-RESTful==0.3.10

# WebSocket支持
Flask-SocketIO==5.3.6
python-socketio==5.9.0
gevent-websocket==0.10.1
gevent==23.9.1
eventlet==0.33.3

# 数据库
SQLAlchemy==2.0.23
PyMySQL==1.1.0

# 缓存
redis==5.0.1

# 异步任务队列
celery==5.3.4

# 安全
Werkzeug==3.0.1

# 验证和序列化
marshmallow==3.20.1
email-validator==2.1.0

# 环境变量
python-dotenv==1.0.0

# 文档解析库
pdfplumber==0.9.0
PyPDF2==3.0.1
python-docx==0.8.11

# 网页解析
beautifulsoup4==4.12.2

# 实用工具
python-magic==0.4.27
chardet==5.2.0

# HTTP请求
requests==2.31.0
urllib3==2.1.0

# AI和机器学习
openai==1.1.1
numpy==1.24.3

# 数据处理
pandas==2.0.3

# 配置管理
python-decouple==3.8

# 自然语言处理 (可选)
spacy==3.7.2
nltk==3.8.1

# 图像处理
Pillow==10.1.0
pytesseract==0.3.10
opencv-python==4.8.1.78

# 监控和日志
prometheus-client==0.19.0

# 测试
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0

# 部署
gunicorn==21.2.0

# 语音转录相关依赖
SpeechRecognition==3.10.0
pyaudio==0.2.11
pydub==0.25.1
openai-whisper==20231117

# 音频处理
librosa==0.10.1
soundfile==0.12.1 